# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moneykeeper",
    "version": "0.1.0-foundation",
    "description": "Moneykeeper context-first contract mounted by the default runtime at unversioned paths."
  },
  "paths": {
...
    "/transactions": {
      "post": {
        "operationId": "recordManualTransaction", "summary": "Record income or expense",
        "description": "Records positive catalog-validated Money through the closed manual income or expense journal recipe.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ManualTransaction"}}}},
        "responses": {"201": {"description": "Transaction posted", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      },
      "get": {
        "operationId": "listLedgerTransactions", "summary": "List transactions",
        "description": "Lists tenant journal facts with stable reverse chronological cursor pagination.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/AfterOccurredAt"},{"$ref":"#/components/parameters/AfterSequence"},{"$ref":"#/components/parameters/ActivityLimit"}], "responses": {"200": {"description": "Transactions", "content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Journal"}}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
...
    "/transactions/{id}/annotation": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "patch": {
        "operationId": "updateTransactionAnnotation", "summary": "Update transaction annotation",
        "description": "Updates mutable transaction metadata at expected_version while keeping the journal and postings immutable.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnnotationUpdate"}}}},
        "responses": {"200": {"description": "Annotation updated", "content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnnotationResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
...
    "/reconciliations/{id}/approve": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "post": {
        "operationId": "approveReconciliation", "summary": "Approve a balance reconciliation",
        "description": "Posts a visible correction only when both case expected_version and captured expected_balance_version still match.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApproveReconciliation"}}}},
        "responses": {"200": {"description": "Reconciliation approved"}, "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
    "/reconciliations/{id}/dismiss": {
      "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
      "post": {
        "operationId": "dismissReconciliation", "summary": "Dismiss a balance reconciliation",
        "description": "Dismisses one pending case at expected_version without creating any journal or changing a balance.",
        "security": [{"bearerAuth": []}], "parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DismissReconciliation"}}}},
        "responses": {"200": {"description": "Reconciliation dismissed"}, "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}, "500": {"$ref": "#/components/responses/InternalServerError"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdempotencyKey": {"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":200}},
      "AfterOccurredAt": {"name":"after_occurred_at","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},
      "AfterSequence": {"name":"after_sequence","in":"query","required":false,"schema":{"type":"integer","format":"int64","minimum":1}},
      "ActivityLimit": {"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":200,"default":50}}
    },
    "responses": {
      "BadRequest": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "NotFound": {"description": "Resource not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Conflict": {"description": "Optimistic version or uniqueness conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "InternalServerError": {"description": "Supporting context unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "MoneyInput": {
        "type": "object", "additionalProperties": false, "required": ["amount", "currency"],
        "properties": {
          "amount": {"type": "string", "pattern": "^-?[0-9]+(\\.[0-9]+)?$"},
          "currency": {"type": "string", "pattern": "^[A-Z]{3}$"}
        }
      },
      "DecimalString": {"type":"string","pattern":"^-?[0-9]+(\\.[0-9]+)?$"},
...
      "AccountKind": {"type":"string","enum":["cash","debit_card","credit_card","current","savings","jar","loan_payable","loan_receivable","system"]},
      "AccountNature": {"type":"string","enum":["asset","liability","equity","income","expense"]},
      "AccountAuthority": {"type":"string","enum":["manual","provider_observed","system"]},
...
      "AccountEffect": {"type":"object","required":["account_id","currency","signed_amount","display_effect","signed_balance","display_balance","balance_version"],"properties":{"account_id":{"type":"string","format":"uuid"},"currency":{"type":"string","pattern":"^[A-Z]{3}$"},"signed_amount":{"$ref":"#/components/schemas/DecimalString"},"display_effect":{"$ref":"#/components/schemas/DecimalString"},"signed_balance":{"$ref":"#/components/schemas/DecimalString"},"display_balance":{"$ref":"#/components/schemas/DecimalString"},"balance_version":{"type":"integer","minimum":0}}},
      "Posting": {"type":"object","required":["id","account_id","account_kind","account_nature","account_authority","position","currency","signed_amount","display_effect"],"properties":{"id":{"type":"string","format":"uuid"},"account_id":{"type":"string","format":"uuid"},"account_kind":{"$ref":"#/components/schemas/AccountKind"},"account_nature":{"$ref":"#/components/schemas/AccountNature"},"account_authority":{"$ref":"#/components/schemas/AccountAuthority"},"position":{"type":"integer","minimum":0},"currency":{"type":"string","pattern":"^[A-Z]{3}$"},"signed_amount":{"$ref":"#/components/schemas/DecimalString"},"display_effect":{"$ref":"#/components/schemas/DecimalString"}}},
      "JournalAnnotation": {"type":"object","required":["version","description","category_id","note","tags","budget_visibility","created_at","updated_at"],"properties":{"version":{"type":"integer","minimum":1},"description":{"type":"string"},"category_id":{"type":["string","null"],"format":"uuid"},"note":{"type":["string","null"]},"tags":{"type":"array","items":{"type":"string"}},"budget_visibility":{"type":"string","enum":["included","excluded"]},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}},
      "JournalRelations": {"type":"object","required":["reverses_transaction_id","corrects_transaction_id","replaces_transaction_id"],"properties":{"reverses_transaction_id":{"type":["string","null"],"format":"uuid"},"corrects_transaction_id":{"type":["string","null"],"format":"uuid"},"replaces_transaction_id":{"type":["string","null"],"format":"uuid"}}},
      "JournalActor": {"type":"object","required":["kind"],"properties":{"kind":{"type":"string","enum":["user","system","external"]},"user_id":{"type":"string","format":"uuid"},"source_kind":{"type":"string"},"source_reference":{"type":"string"}}},
      "Correction": {"type":"object","required":["account_id","before_display_balance","target_display_balance","display_delta","observed_balance_version","reason","observed_at"],"properties":{"account_id":{"type":"string","format":"uuid"},"before_display_balance":{"$ref":"#/components/schemas/DecimalString"},"target_display_balance":{"$ref":"#/components/schemas/DecimalString"},"display_delta":{"$ref":"#/components/schemas/DecimalString"},"observed_balance_version":{"type":"integer","minimum":0},"reason":{"type":"string"},"observed_at":{"type":"string","format":"date-time"}}},
      "Journal": {"type":"object","required":["id","user_id","ledger_sequence","source","purpose","actor","description","occurred_at","recorded_at","correlation_id","relations","postings","annotation","reversed_by_journal_id","replaced_by_journal_id","correction"],"properties":{"id":{"type":"string","format":"uuid"},"user_id":{"type":"string","format":"uuid"},"ledger_sequence":{"type":"integer","format":"int64","minimum":1},"source":{"type":"string","enum":["manual","import","system","correction","reconciliation"]},"purpose":{"type":"string","enum":["ordinary","correction","reversal","approved_reconciliation"]},"actor":{"$ref":"#/components/schemas/JournalActor"},"description":{"type":"string"},"occurred_at":{"type":"string","format":"date-time"},"recorded_at":{"type":"string","format":"date-time"},"correlation_id":{"type":"string","format":"uuid"},"relations":{"$ref":"#/components/schemas/JournalRelations"},"postings":{"type":"array","minItems":2,"items":{"$ref":"#/components/schemas/Posting"}},"annotation":{"oneOf":[{"$ref":"#/components/schemas/JournalAnnotation"},{"type":"null"}]},"reversed_by_journal_id":{"type":["string","null"],"format":"uuid"},"replaced_by_journal_id":{"type":["string","null"],"format":"uuid"},"correction":{"oneOf":[{"$ref":"#/components/schemas/Correction"},{"type":"null"}]}}},
      "TransactionResult": {"type":"object","required":["journal_entry_id","effects","annotation_version","replayed"],"properties":{"journal_entry_id":{"type":"string","format":"uuid"},"effects":{"type":"array","items":{"$ref":"#/components/schemas/AccountEffect"}},"annotation_version":{"type":"integer","minimum":1},"replayed":{"type":"boolean"}}},
      "AnnotationResult": {"type":"object","required":["journal_entry_id","version","replayed"],"properties":{"journal_entry_id":{"type":"string","format":"uuid"},"version":{"type":"integer","minimum":1},"replayed":{"type":"boolean"}}},
...
      "ManualTransaction": {
        "type": "object", "additionalProperties": false,
        "required": ["account_id", "kind", "amount", "description"],
        "properties": {
          "account_id": {"type": "string", "format": "uuid"}, "kind": {"type": "string", "enum": ["income", "expense"]},
          "amount": {"$ref": "#/components/schemas/MoneyInput"}, "description": {"type": "string"},
          "category_id": {"type": ["string", "null"], "format": "uuid"}, "note":{"type":["string","null"]}, "tags": {"type": "array", "items": {"type": "string"}},
          "budget_visibility":{"type":"string","enum":["included","excluded"],"default":"included"},"occurred_at":{"type":"string","format":"date-time"}
        }
      },
      "AnnotationUpdate": {
        "type": "object", "additionalProperties": false, "required": ["expected_version"],
        "properties": {"expected_version": {"type": "integer", "minimum": 1}, "description": {"type": "string"}, "category_id": {"type": ["string", "null"], "format": "uuid"},"clear_category":{"type":"boolean","default":false},"note":{"type":["string","null"]},"clear_note":{"type":"boolean","default":false},"tags":{"type":"array","items":{"type":"string"}},"budget_visibility":{"type":"string","enum":["included","excluded"]},"occurred_at":{"type":"string","format":"date-time"}}
      },
...
      "ApproveReconciliation": {
        "type": "object", "additionalProperties": false,
        "required": ["expected_version", "expected_balance_version", "reason"],
        "properties": {"expected_version": {"type": "integer", "minimum": 1}, "expected_balance_version": {"type": "integer", "minimum": 1}, "reason": {"type": "string"},"occurred_at":{"type":"string","format":"date-time"}}
      },
      "DismissReconciliation": {
        "type": "object", "additionalProperties": false, "required": ["expected_version", "reason"],
        "properties": {"expected_version": {"type": "integer", "minimum": 1}, "reason": {"type": "string"},"occurred_at":{"type":"string","format":"date-time"}}
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {"error": {"type": "string"}}
      }
    }
  }
}